Add Easy Java Solution for 69. Sqrt(x) from LeetCode

+++ LeetCode/Patterns/Binary Search/69. Sqrtx/Solution.java
@@ -1,19 +1,19 @@
 class Solution {
     public int mySqrt(int x) {
 
-       int sqrt = 0;
+       long sqrt = 0;
         while(true){
         
             if(sqrt*sqrt < x){
                 sqrt++;
                 if(sqrt*sqrt > x){
-                    return sqrt - 1 ;
+                    return (int) sqrt - 1 ;
                 }
             }
             else{
                 break;
             }
         }
-        return sqrt;
+        return (int) sqrt;
     }
 }
